Update community and developer chat locations
As part of CNCF, we have new Slack based channels where we can be
reached.  The `podman` channel is for the community and focuses on how
to run and use Podman.  The `podman-dev` is for the development of
podman and its related projects like buildah and skopeo.

Given recent community conversations around our community channels, I
also removed the references to IRC given it has low volume and there are
a slew of alternatives.

Signed-off-by: Brent Baude <bbaude@redhat.com>